Develop login form with Formik and Yup
The Formik library allows for forms to be built in React with
greater ease: managing state, form submissions, and validation.
Yup provides an easy way to create a validation schema, which
can then be fed into Formik. Basic functionality has been built,
and the next step is refactoring for more simplistic code.