Implement flattened store navigation data structure
A storeNavigationFlat.json file was created, flattening the nesting
of the previous storeNavigation file, and making the parent slugs
held in the rootUrl field. This is test data and is experimenting
with the best way to fetch data from the API. This file was also
implemented in gatsby node, successfully creating all the required
pages with their appropriate urls.