Add ability to search multiple items
A function within the recipes page allows for a 2-d array
of search entries to be created and parsed to determine
which items are common between all the search terms. For instance,
if you search 'chicken', 'rice', and 'teriyaki', you'll be
returned every recipe that has those words in its title.